requirements.txt
streamlit
tensorflow-cpu==2.15.0
Pillow
numpy
pandas
huggingface_hub
